# skills/kanban-openapi.yaml
openapi: 3.0.0
info:
  title: AI Marketing Crew Agent API
  description: API for AI Agents to interact with the AI Marketing Crew collaboration board. Enables task creation, state transition, human-in-the-loop requests, and task retrieval.
  version: 1.0.0
servers:
  - url: https://amc-kanban.immedi.ai/api
    description: Production Server
  - url: http://localhost:3000/api
    description: Local Development Server
paths:
...
  /tasks/{id}:
    get:
      summary: Get task details
      operationId: getTaskDetails
      description: Fetch full details of a task including its current description and materials.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          description: Bearer token received from your Initialization Instruction
          schema:
            type: string
      responses:
        '200':
          description: Task details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
                
    patch:
      summary: Update task details
      operationId: updateTaskDetails
      description: Update the content, title, or materials of an existing task. Use this to log progress in the description.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          description: Bearer token received from your Initialization Instruction
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                materials:
                  type: string
                brandId:
                  type: string
                  nullable: true
                  description: Brand ID this task belongs to. Pass null to clear legacy unscoped tasks.
                assigneeId:
                  type: string
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{id}/status:
    patch:
      summary: Update task status or request human input
      operationId: updateTaskStatus
      description: Move the task between columns. Crucially, set status to "pending" and provide "requiredInput" to ask human operators for help.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          description: Bearer token received from your Initialization Instruction
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [todo, in_progress, pending, done, void]
                requiredInput:
                  type: string
                  description: Only used when status is "pending". Explains exactly what the AI needs from the human to proceed. Set to null when clearing.
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
...
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        materials:
          type: string
          nullable: true
        weight:
          type: integer
        requiredInput:
          type: string
          nullable: true
        assigneeId:
          type: string
          nullable: true
        brandId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...